Refactoring. Removed all the O(n) and O(n^2) searches and replaced with O(1) lookups. Also moved to a priority queue approach for Dijkstra. Now runs in about a second not 20 minutes

+++ 2021/15/context.hpp
@@ -9,12 +9,12 @@ namespace aoc
     struct node {
         int x;
         int y;
-        int id;
+        long id;
     };
 
     struct edge {
-        int source_id;
-        int destination_id;
+        long source_id;
+        long destination_id;
         int weight;
     };
